@@ -1,9 +1,10 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import React from "react";
+import {createRoot} from "react-dom/client";
+import App from "./App";
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
+it("renders without crashing", () => {
+  const div = document.createElement("div");
+  const root = createRoot(div); // createRoot(div!) if you use TypeScript
+  root.render(<App />);
+  root.unmount();
 });
