@@ -1,8 +1,8 @@
-import {BrowserRouter, Redirect, Route, Switch} from "react-router-dom";
-import {GameGuard} from "components/routing/routeProtectors/GameGuard";
-import GameRouter from "components/routing/routers/GameRouter";
-import {LoginGuard} from "components/routing/routeProtectors/LoginGuard";
-import Login from "components/views/Login";
+import {BrowserRouter, Navigate, Route, Routes} from "react-router-dom";
+import {GameGuard} from "../routeProtectors/GameGuard";
+import GameRouter from "./GameRouter";
+import {LoginGuard} from "../routeProtectors/LoginGuard";
+import Login from "../../views/Login";
 
 /**
  * Main router of your application.
@@ -16,21 +16,21 @@ import Login from "components/views/Login";
 const AppRouter = () => {
   return (
     <BrowserRouter>
-      <Switch>
-        <Route path="/game">
-          <GameGuard>
-            <GameRouter base="/game"/>
-          </GameGuard>
+      <Routes>
+
+        <Route path="/game/*" element={<GameGuard />}>
+            <Route path="/game/*" element={<GameRouter base="/game"/>} />
         </Route>
-        <Route exact path="/login">
-          <LoginGuard>
-            <Login/>
-          </LoginGuard>
+
+        <Route path="/login" element={<LoginGuard />}>
+            <Route path="/login" element={<Login/>} />
         </Route>
-        <Route exact path="/">
-          <Redirect to="/game"/>
-        </Route>
-      </Switch>
+
+        <Route path="/" element={
+          <Navigate to="/game" replace />
+        }/>
+
+      </Routes>
     </BrowserRouter>
   );
 };
