@@ -1,7 +1,7 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "styles/index.scss";
-import App from "App";
+import {createRoot} from "react-dom/client";
+import "./styles/index.scss";
+import App from "./App";
 
 /**
  * This is the entry point of your React application where the root element is in the public/index.html.
@@ -9,4 +9,6 @@ import App from "App";
  * Applications built with just React usually have a single root DOM node.
  * More: https://reactjs.org/docs/rendering-elements.html
  */
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById("app");
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App tab="home" />);
