@@ -1,11 +1,11 @@
-import React from "react";
-import Header from "components/views/Header";
-import AppRouter from "components/routing/routers/AppRouter";
+import Header from "./components/views/Header";
+import AppRouter from "./components/routing/routers/AppRouter";
 
 /**
  * Happy coding!
  * React Template by Lucas Pelloni
  * Overhauled by Kyrill Hux
+ * Updated by Marco Leder
  */
 const App = () => {
     return (
