@@ -1,8 +1,8 @@
-import 'styles/ui/BaseContainer.scss';
+import "../../styles/ui/BaseContainer.scss";
 import PropTypes from "prop-types";
 
 const BaseContainer = props => (
-  <div {...props} className={`base-container ${props.className ?? ''}`}>
+  <div {...props} className={`base-container ${props.className ?? ""}`}>
     {props.children}
   </div>
 );
