@@ -1,5 +1,5 @@
 import React from "react";
-import "styles/ui/Spinner.scss";
+import "../../styles/ui/Spinner.scss";
 
 export const Spinner = () => (
   <div className="loading-spinner">
