@@ -1,7 +1,7 @@
 import React from "react";
-import { ReactLogo } from "components/ui/ReactLogo";
+import {ReactLogo} from "../ui/ReactLogo";
 import PropTypes from "prop-types";
-import "styles/views/Header.scss";
+import "../../styles/views/Header.scss";
 
 /**
  * This is an example of a Functional and stateless component (View) in React. Functional components are not classes and thus don't handle internal state changes.
