@@ -1,4 +1,4 @@
-import { isProduction } from 'helpers/isProduction'
+import { isProduction } from "./isProduction"
 
 /**
  * This helper function returns the current domain of the API.
@@ -7,8 +7,8 @@ import { isProduction } from 'helpers/isProduction'
  * @returns {string}
  */
 export const getDomain = () => {
-  const prodUrl = 'https://my-server-url.oa.r.appspot.com/' // TODO: insert your prod url for server (once deployed)
-  const devUrl = 'http://localhost:8080'
+  const prodUrl = "https://my-server-url.oa.r.appspot.com/" // TODO: insert your prod url for server (once deployed)
+  const devUrl = "http://localhost:8080"
 
   return isProduction() ? prodUrl : devUrl
 }
